fish_prompt: Make users aware they're in an SSH session
We can easily inform users they're in an SSH session by displaying the
'user@hostname' string as part of the prompt string.

The way chosen to do this here is checking if either "$SSH_CLIENT" or
"$SSH_TTY" are set and acting accordingly to whether they're set or not.

The proposed changes do not change any of the colours, but there's a
noticeable difference on the prompt itself, as shown below:
* Before this patch:
⋊> ~

* After this patch:
⋊> fidencio@pessoa ⋊> ~

It's important to make it very much clear that this difference only
happens when the user is connected to a fish terminal via SSH, and users
can opt-out from this behaviour by doing:
```sh
set -g theme_ignore_ssh_awareness yes
```

Fixes: #22

Signed-off-by: Fabiano Fidêncio <fabiano@fidencio.org>