fish_right_prompt: Make it optional
Allow users to easily decide whether or not the `fish_right_prompt`
should be displayed.

The current way to do so is creating an empty `fish_right_prompt`
function in `~/.config/fish/functions/fish_mode_prompt.fish`, which
although is very well documented and the default way to do so, is less
handy than having an option on the theme to do so.

Fixes: #24

Signed-off-by: Fabiano Fidêncio <fabiano@fidencio.org>